requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
Django==3.0.8
django-cors-headers==3.4.0
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.0
dnspython==2.0.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mongoengine==0.20.0
pycodestyle==2.6.0
pylint==2.5.3
pymongo==3.11.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
